move utils.py, add timeout arg, remove desc
